LayoutTests/css3/filters/custom-with-at-rule-syntax/parsing-at-rule-valid-expected.txt

Test at-rule parsing for @-webkit-filter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



========================================
Filter at-rule tests.
========================================

Empty rule, separated by single spaces.
@-webkit-filter my-filter { }
PASS cssRule.cssText is "@-webkit-filter my-filter { }"
PASS cssRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true
PASS declaration.length is 0

Empty rule, separated by multiple spaces.
   @-webkit-filter   my-filter   {   }   
PASS cssRule.cssText is "@-webkit-filter my-filter { }"
PASS cssRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true
PASS declaration.length is 0

Empty rule, no extra whitespace.
@-webkit-filter my-filter{}
PASS cssRule.cssText is "@-webkit-filter my-filter { }"
PASS cssRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true
PASS declaration.length is 0

Rule with arbitrary properties.
@-webkit-filter my-filter { width: 100px; height: 100px; }
PASS cssRule.cssText is "@-webkit-filter my-filter { width: 100px; height: 100px; }"
PASS cssRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true
PASS declaration.getPropertyValue('width') is "100px"
PASS declaration.getPropertyValue('height') is "100px"
PASS declaration.length is 2

========================================
Nested filter at-rule tests.
========================================

Nested rule.
@-webkit-filter parent-filter { @-webkit-filter child-filter { } }
PASS cssRule.cssText is "@-webkit-filter parent-filter { }"
PASS cssRule.type is 17
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true

Twice nested rule.
@-webkit-filter parent-filter { @-webkit-filter child-filter { @-webkit-filter grandchild-filter } }
PASS cssRule.cssText is "@-webkit-filter parent-filter { }"
PASS cssRule.type is 17
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true

Nested rule inside arbitrary rule.
@font-face { @-webkit-filter child-filter { } }
PASS cssRule.cssText is "@font-face { }"
PASS cssRule.type is 5
PASS cssRule instanceof CSSFontFaceRule is true
PASS cssRule.constructor === CSSFontFaceRule is true
PASS cssRule.__proto__ === CSSFontFaceRule.prototype is true

Nested rule inside media query.
@media screen { @-webkit-filter child-filter { } }
PASS cssRule.cssText is "@media screen { \n  @-webkit-filter child-filter { }\n}"
PASS cssRule.type is 4
PASS cssRule instanceof CSSMediaRule is true
PASS cssRule.constructor === CSSMediaRule is true
PASS cssRule.__proto__ === CSSMediaRule.prototype is true
PASS cssRule.cssText is "@-webkit-filter child-filter { }"
PASS cssRule.type is 17
PASS cssRule instanceof WebKitCSSFilterRule is true
PASS cssRule.constructor === WebKitCSSFilterRule is true
PASS cssRule.__proto__ === WebKitCSSFilterRule.prototype is true
PASS successfullyParsed is true

TEST COMPLETE
